api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
deprecation==2.1.0
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.14
packaging==23.2
postgrest==0.15.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
